+++ internal/model/cart_product_for_order.go
@@ -5,7 +5,7 @@ import "time"
 type CartProductForOrder struct {
 	Title        string
 	Price        float32
-	Quantity     uint16
+	Quantity     uint32
 	Image        string
 	Weight       float32
 	URL          string
