+++ internal/model/order_products.go
@@ -9,9 +9,9 @@ type ProductOrder struct {
 	ID           uint32
 	ProductID    uint32
 	OptionID     *uint32
-	Count        uint16
+	Count        uint32
 	Weight       float32
-	Cost         uint16
+	Cost         uint32
 	OrderID      uuid.UUID
 	DeliveryDate time.Time
 	Name         string
